Added tests

+++ databallpy/load_data/tracking_data/_add_player_data_to_dict.py
@@ -23,8 +23,8 @@ def _add_player_data_to_dict(team_id:str, shirt_num:str, x:str, y:str, data: dic
     if f"{team}_{shirt_num}_x" not in data.keys():  # create keys for new player
         data[f"{team}_{shirt_num}_x"] = [np.nan] * len(data["timestamp"])
         data[f"{team}_{shirt_num}_y"] = [np.nan] * len(data["timestamp"])
-
-    data[f"{team}_{shirt_num}_x"][idx] = int(x)
-    data[f"{team}_{shirt_num}_y"][idx] = int(y)
+    #import pdb;pdb.set_trace()
+    data[f"{team}_{shirt_num}_x"][idx] = float(x)
+    data[f"{team}_{shirt_num}_y"][idx] = float(y)
 
     return data
